added header

+++ drivers/mrf24j40/include/mrf24j40_internal.h
@@ -47,7 +47,7 @@ void mrf24j40_reg_write_short(const mrf24j40_t *dev, const uint8_t addr, const u
  * @return the value of the specified register
  */
  
-uint8_t mrf24j40_reg_read_long(const mrf24j40_t *dev, const uint8_t addr);
+uint8_t mrf24j40_reg_read_long(const mrf24j40_t *dev, const uint16_t addr);
 
 /**
  * @brief Write to a register at address `addr` from device `dev`. Register with 10bit address
@@ -57,4 +57,4 @@ uint8_t mrf24j40_reg_read_long(const mrf24j40_t *dev, const uint8_t addr);
  * @param[in] value value to write to the given register
  */
  
-void mrf24j40_reg_write_short_long(const mrf24j40_t *dev, const uint8_t addr, const uint8_t value);
+void mrf24j40_reg_write_short_long(const mrf24j40_t *dev, const uint16_t addr, const uint8_t value);
